move instruction-ai into o-mate bundle
