--- extensions/instruction-ai/package.json
@@ -1,42 +1,0 @@
-{
-	"name": "instruction-ai",
-	"description": "Please enter a description for your extension",
-	"icon": "extension",
-	"version": "1.0.0",
-	"keywords": [
-		"directus",
-		"directus-extension",
-		"directus-extension-operation"
-	],
-	"type": "module",
-	"files": [
-		"dist"
-	],
-	"directus:extension": {
-		"type": "operation",
-		"path": {
-			"app": "dist/app.js",
-			"api": "dist/api.js"
-		},
-		"source": {
-			"app": "src/app.ts",
-			"api": "src/api.ts"
-		},
-		"host": "^10.10.0"
-	},
-	"scripts": {
-		"build": "directus-extension build",
-		"dev": "directus-extension build -w --no-minify",
-		"link": "directus-extension link",
-		"validate": "directus-extension validate"
-	},
-	"devDependencies": {
-		"@directus/extensions-sdk": "13.1.0",
-		"@types/node": "^22.15.3",
-		"typescript": "^5.8.3",
-		"vue": "^3.5.13"
-	},
-	"dependencies": {
-		"openai": "^4.96.2"
-	}
-}
