+++ extensions/o-mate/src/instruction-ai/src/app.ts
@@ -1,0 +1,15 @@
+import { defineOperationApp } from '@directus/extensions-sdk';
+
+export default defineOperationApp({
+	id: 'o-mate-instruction-ai',
+	name: 'Instruction AI',
+	icon: 'signpost',
+	description: 'Get structured data from unstructured instructions using AI.',
+	overview: ({ text }) => [
+		{
+			label: 'Text',
+			text: text,
+		},
+	],
+	options: [],
+});
